Improved unit tests for WA progress.
